Add the comment, post, delete and the logged in user profile picture in the home.ejs I added the delete, comment form and a check on the logged in user so that the profile image can be added and displayed I added a post request to the deletePost function in the homeController I edited the delete form to include the same method = post and the same url with what we have in the homeRouter so that the delete functionality can work While trying to access the post Id in the deletePost function in the homeController, I included the user because of the relation between them I created the comment.ejs file Changes: The home and post Router were edited, the home.ejs and comment.ejs were edited, prisma schema was updated to include the onDelete cascade so that comments can be deleted alongside the post The home, post and comment Controllers were edited

+++ odinbook/controllers/postController.js
@@ -30,6 +30,32 @@ exports.createPost = async (req, res) => {
   }
 };
 
+exports.createPost = async (req, res) => {
+  try {
+    if (!req.user) {
+      return res.status(401).render("../odinbook/views/error", { error: 'You must be logged in to create a post' });
+    }
+
+    const { content } = req.body;
+    const userId = req.user.id;
+    const image = req.file ? req.file.filename : null;
+
+    const post = await prisma.post.create({
+      data: {
+        content,
+        image,
+        user: { connect: { id: userId } },
+      },
+    });
+
+    console.log(post);
+    res.render("../odinbook/views/post", { post });
+  } catch (err) {
+    console.error(err);
+    res.status(500).render("../odinbook/views/error", { error: 'Failed to create post' });
+  }
+};
+
 
 
 // Get all posts
@@ -86,15 +112,32 @@ exports.updatePost = async (req, res) => {
 // Delete a post
 exports.deletePost = async (req, res) => {
   try {
+    if (!req.user) {
+      return res.status(401).json({ message: 'You must be logged in to delete a post' });
+    }
+
     const { id } = req.params;
+    const post = await prisma.post.findUnique({
+      where: { id },
+      include: { user: true },
+    });
+
+    if (!post) {
+      return res.status(404).json({ message: 'Post not found' });
+    }
+
+    if (post.user.id !== req.user.id) {
+      return res.status(403).json({ message: 'You do not have permission to delete this post' });
+    }
+
     await prisma.post.delete({
       where: { id },
     });
-    res.redirect("/posts");
+
+    res.redirect("/home");
   } catch (err) {
     console.error(err);
     res.status(500).render("../odinbook/views/error", { error: 'Failed to delete post' });
   }
 };
-
